+++ src/render/Material.h
@@ -6,6 +6,7 @@
  */
 #pragma once
 
+#include "core/EntityMaterial.h"
 #include "render/SamplerState.h"
 #include "render/Texture2D.h"
 
@@ -18,9 +19,12 @@ namespace lrender {
 struct Material {
     std::string name{"Default"};
     DirectX::SimpleMath::Color baseColorFactor{1.0F, 1.0F, 1.0F, 1.0F};
+    float diffuseStrength{1.0F};
+    DirectX::SimpleMath::Color specularColor{1.0F, 1.0F, 1.0F, 1.0F};
     float specularStrength{0.25F};
     float shininess{32.0F};
     bool doubleSided{false};
+    SurfaceDisplayMode displayMode{SurfaceDisplayMode::LitTextured};
     std::shared_ptr<Texture2D> baseColorTexture;
     std::shared_ptr<SamplerState> sampler;
 };